feat: better reporting of build log
With the changes I have PR’ed to Parcel, you can override the log stream and provide your own stream to write messages to.

This allows the plugin to hook into these events, parse and format them a little, and generate a much better progress UI that should let the user know what is going on in the build.